Added base implementation of regression model, data & normalised data
Important notice: the implementation uploaded (Kalman_Filter_Model) has an issue with the rolling alpha and rolling beta data not showing beyond 05/2019. The cause of this is speculated to be over inflated covariance matrices but this is not concrete yet. Important to explore more ideas behind the potential cause of this issue.